requirements.txt
flask
flask-socketio
flask-login
psycopg2-binary
requests
reportlab
google-generativeai
python-dotenv
